move strptime validation logic into ctimefmt package
