+++ internal/coreinternal/timeutils/internal/ctimefmt/ctimefmt_test.go
@@ -10,6 +10,7 @@
 package ctimefmt
 
 import (
+	"github.com/stretchr/testify/require"
 	"testing"
 	"time"
 )
@@ -72,6 +73,57 @@ func TestZulu(t *testing.T) {
 				// The former returns a Time with the UTC timezone, the latter returns a Time with a 0000 time zone offset.
 				// (See Go's documentation for `time.Parse`.)
 				t.Errorf("Given: %v, expected: %v", dt, dt1)
+			}
+		})
+	}
+}
+
+func TestValidate(t *testing.T) {
+	type args struct {
+		layout string
+	}
+	tests := []struct {
+		name    string
+		args    args
+		wantErr string
+	}{
+		{
+			name: "valid format",
+			args: args{
+				layout: "%Y-%m-%d %H:%M:%S.%f",
+			},
+			wantErr: "",
+		},
+		{
+			name: "invalid fractional second",
+			args: args{
+				layout: "%Y-%m-%d-%H-%M-%S:%L",
+			},
+			wantErr: "invalid fractional seconds directive: ':%L'. must be preceded with '.' or ','",
+		},
+		{
+			name: "format including decimal",
+			args: args{
+				layout: "2006-%m-%d-%H-%M-%S:%L",
+			},
+			wantErr: "format string should not contain decimals",
+		},
+		{
+			name: "unsupported directive",
+			args: args{
+				layout: "%C-%m-%d-%H-%M-%S.%L",
+			},
+			wantErr: "invalid strptime format: [unsupported ctimefmt.ToNative() directive: %C]",
+		},
+	}
+	for _, tt := range tests {
+		t.Run(tt.name, func(t *testing.T) {
+			err := Validate(tt.args.layout)
+
+			if tt.wantErr != "" {
+				require.ErrorContains(t, err, tt.wantErr)
+			} else {
+				require.NoError(t, err)
 			}
 		})
 	}
