+++ internal/coreinternal/timeutils/parser_test.go
@@ -62,50 +62,6 @@ func Test_setTimestampYear(t *testing.T) {
 	})
 }
 
-func TestValidateStrptime(t *testing.T) {
-	type args struct {
-		layout string
-	}
-	tests := []struct {
-		name    string
-		args    args
-		wantErr string
-	}{
-		{
-			name: "valid format",
-			args: args{
-				layout: "%Y-%m-%d %H:%M:%S.%f",
-			},
-			wantErr: "",
-		},
-		{
-			name: "invalid fractional second",
-			args: args{
-				layout: "%Y-%m-%d-%H-%M-%S:%L",
-			},
-			wantErr: "invalid fractional seconds directive: ':%L'. must be preceded with '.' or ','",
-		},
-		{
-			name: "format including decimal",
-			args: args{
-				layout: "2006-%m-%d-%H-%M-%S:%L",
-			},
-			wantErr: "format string should not contain decimals",
-		},
-	}
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			err := ValidateStrptime(tt.args.layout)
-
-			if tt.wantErr != "" {
-				require.ErrorContains(t, err, tt.wantErr)
-			} else {
-				require.NoError(t, err)
-			}
-		})
-	}
-}
-
 func TestValidateGotime(t *testing.T) {
 	type args struct {
 		layout string
